Revert "PREMIUMAPP-3461: Remove lazy_static, remove unsafe blocks, split "interface" file into smaller modules. (#82)"
This reverts commit 7c61c81e83b86a4cb156702152a6c4fbc22aeaa3.
