+++ src/device/rdk.rs
@@ -2,10 +2,9 @@ pub mod applications;
 pub mod device;
 pub mod health_check;
 pub mod input;
-pub mod configuration;
+pub mod interface;
 pub mod operations;
 pub mod output;
 pub mod system;
 pub mod version;
 pub mod voice;
-pub mod interface;
