+++ src/device/rdk/input/key_press.rs
@@ -1,8 +1,8 @@
 use crate::dab::structs::DabError;
 use crate::dab::structs::KeyPressRequest;
 use crate::dab::structs::KeyPressResponse;
-use crate::device::rdk::configuration::get_keycode;
-use crate::device::rdk::interface::http::http_post;
+use crate::device::rdk::interface::get_keycode;
+use crate::device::rdk::interface::http_post;
 use serde::{Deserialize, Serialize};
 
 #[allow(non_snake_case)]
